+++ src/main/java/org/esgi/el_presidente/core/ressources/IslandPart.java
@@ -1,9 +1,9 @@
 package org.esgi.el_presidente.core.ressources;
 
 public interface IslandPart {
-  void shrink(double deductSize);
+  void shrink(int deductSize);
 
-  void expand(double additionalSize);
+  void expand(int additionalSize);
 
-  double getSize();
+  int getSize();
 }
